Reset to JWT use only on sign in
Was implemented to provide a session for guests.
But since this can be implemented via SessionStore
or LocalStore, it won't be necessary any more.

Another problem coult be a custom Captcha. But in
that case, the Captcha can generate an encrypted
solution string, which changes every hour or so and
which can be send to the client. IF anyone tried to
crack that string, a new one has already been
generated and the secret behind that encryption
has already changed.